Refactor batch_reconstruction & reed_solomon, move code into utils.py
This commit began as a bugfix for inducing faults in reed_solomon.py,
but ended up refactoring code to be better stylistically as I went
along. I moved shared utility functions into utils.py, such as the
type checker and generic methods from batch_reconstruction.

I also added in functionality to the type checker, so now it supports
arrays of types, as well as constraints such as 'callable(func)'.